Converted double in PID to float

+++ stm32/Firmware/app/include/app/pid.h
@@ -33,7 +33,7 @@ namespace app {
          * @param ki Integral gain.
          * @param kd Derivative gain.
          */
-        PIDController(double kp, double ki, double kd);
+        PIDController(float kp, float ki, float kd);
 
         /**
          * @brief Construct with gains and explicit output/integral limits.
@@ -45,8 +45,8 @@ namespace app {
          * @param min_integral Minimum integral accumulator (anti-windup).
          * @param max_integral Maximum integral accumulator (anti-windup).
          */
-        PIDController(double kp, double ki, double kd, double min_out, double max_out,
-                      double min_integral, double max_integral);
+        PIDController(float kp, float ki, float kd, float min_out, float max_out,
+                      float min_integral, float max_integral);
 
         /**
          * @brief Set the PID gains.
@@ -55,7 +55,7 @@ namespace app {
          * @param kd Derivative gain.
          * @return true on success, false if any gain is non-finite.
          */
-        bool set_gains(double kp, double ki, double kd);
+        bool set_gains(float kp, float ki, float kd);
 
         /**
          * @brief Configure the desired target (setpoint) used to compute error.
@@ -66,13 +66,13 @@ namespace app {
          *       term will not produce a spike on the immediately following
          *       `update()` call; derivative resumes after the next sample.
          */
-        bool set_setpoint(double setpoint);
+        bool set_setpoint(float setpoint);
 
         /**
          * @brief Get the currently configured setpoint.
          * @return The setpoint value.
          */
-        double get_setpoint() const;
+        float get_setpoint() const;
 
         /**
          * @brief Configure output clamping range.
@@ -80,19 +80,19 @@ namespace app {
          * @param max_out Maximum output value.
          * @return true on success, false if inputs are invalid (non-finite or min>max).
          */
-        bool set_output_limits(double min_out, double max_out);
+        bool set_output_limits(float min_out, float max_out);
 
         /**
          * @brief Get minimum output clamp.
          * @return Minimum output value.
          */
-        double get_min_output() const;
+        float get_min_output() const;
 
         /**
          * @brief Get maximum output clamp.
          * @return Maximum output value.
          */
-        double get_max_output() const;
+        float get_max_output() const;
 
         /**
          * @brief Configure integral anti-windup limits.
@@ -100,19 +100,19 @@ namespace app {
          * @param max_integral Maximum integral accumulator.
          * @return true on success, false if inputs are invalid.
          */
-        bool set_integral_limits(double min_integral, double max_integral);
+        bool set_integral_limits(float min_integral, float max_integral);
 
         /**
          * @brief Get minimum integral clamp.
          * @return Minimum integral value.
          */
-        double get_min_integral() const;
+        float get_min_integral() const;
 
         /**
          * @brief Get maximum integral clamp.
          * @return Maximum integral value.
          */
-        double get_max_integral() const;
+        float get_max_integral() const;
 
         /**
          * @brief Set derivative filter time constant (first-order) in seconds.
@@ -120,20 +120,20 @@ namespace app {
          * @param tau_seconds Time constant in seconds.
          * @return true on success, false if tau is negative or non-finite.
          */
-        bool set_derivative_filter(double tau_seconds);
+        bool set_derivative_filter(float tau_seconds);
 
         /**
          * @brief Get configured derivative filter time constant in seconds.
          * @return Time constant (0 means no filtering).
          */
-        double get_derivative_filter() const;
+        float get_derivative_filter() const;
 
         /**
          * @brief Return the last output produced by `update()`.
          * @note Before any `update()` call the value equals clamp(0.0, min_out, max_out).
          * @return Most recent control output.
          */
-        double get_last_output() const;
+        float get_last_output() const;
 
         /**
          * @brief Reset internal integrator, derivative state and previous sample.
@@ -145,37 +145,37 @@ namespace app {
          *        `measurement` and elapsed time `dt` (seconds).
          * @param measurement Current sensor measurement.
          * @param dt Elapsed time in seconds since last update.
-         * @note The derivative term is computed on the measurement
-         *       (d(measurement)/dt) and applied with a negative sign to avoid
-         *       producing spikes when the setpoint changes. This is equivalent
-         *       to computing d(error)/dt when the setpoint is constant but
-         *       prevents derivative 'kick' on setpoint steps.
+         * @note Derivative is computed from the measurement (d(measurement)/dt)
+         *       and negated so that an increasing measurement reduces the output
+         *       when `Kd` is positive. This prevents derivative "kick" on
+         *       immediate setpoint changes while remaining equivalent to
+         *       d(error)/dt when the setpoint is constant.
          * @return Clamped control output.
          */
-        double update(double measurement, double dt);
+        float update(float measurement, float dt);
 
       private:
         // Gain variables
-        double Kp_ {};
-        double Ki_ {};
-        double Kd_ {};
+        float Kp_ {};
+        float Ki_ {};
+        float Kd_ {};
 
         // Target value
-        double setpoint_ {};
-        double min_out_ {std::numeric_limits<double>::lowest()};
-        double max_out_ {std::numeric_limits<double>::max()};
+        float setpoint_ {};
+        float min_out_ {std::numeric_limits<float>::lowest()};
+        float max_out_ {std::numeric_limits<float>::max()};
 
         // Previous raw measurement and flag indicating availability
-        double prev_measurement_ {};
+        float prev_measurement_ {};
         bool has_prev_measurement_ {false};
 
-        double integral_ {};
-        double min_integral_ {std::numeric_limits<double>::lowest()};
-        double max_integral_ {std::numeric_limits<double>::max()};
+        float integral_ {};
+        float min_integral_ {std::numeric_limits<float>::lowest()};
+        float max_integral_ {std::numeric_limits<float>::max()};
 
-        double derivative_filter_tau_ {};
-        double derivative_ {};
-        double last_output_ {};
+        float derivative_filter_tau_ {};
+        float derivative_ {};
+        float last_output_ {};
     };
 } // namespace app
